Dev builds compiling again (univ. binary trick for header-only targets)
Setting CMAKE_OSX_ARCHITECTURES=arm64;x86_64 only seems to work when we have header-only dependencies (which makes sense). For universal binaries of the dev builds, we'd have to compile arm & x86 separately and combine them, but it's not important that these targets be universal binaries so I just set the OS X builds of these to be whatever default cmake likes for the given Mac.