Attempt at using bit-mask to keep track of missing values
It's a messy and disappointing result. Originally tried "boost::dynamic_bitset" so that a variable-sized E can be accommodated, however this introduced huge slowdowns malloc/free calls. The std::bitset<N> was a faster alternative, but it restricts the maximum E we can use in the plugin. There was no significant time saved even for a small N (equiv. maximum E), and the SIMD didn't kick in with the bit masks as I had hoped. Another idea to try is to have std::set to track the missing value indices, but it'd only work well if there weren't many missing vals.