Pull our structs/classes which are shared into common.cpp/h
This is basically the first step in splitting the plugin into the client/server model. This means the client (plugin) can just #include "common.h" instead of #include "edm.h" as it would just need to know the input/output types for EDM but it wouldn't need to have the implementation for the EDM methods.